refactor: organize frontend pages and layout
