+++ server/frontend/src/layout/MainSidebar.tsx
